Added new database functionality through Back4App
I wanted to migrate from heroku because the load up times were atrocious but since my database was hosted on heroku I had to migrate that somewhere else. I found Back4App which is a free api and mongodb hosting service and decided to use that because it was easy to use and I could deprecate my api and just use theirs.